refactor(mf2): Collect files in a new src/data-model/ directory
