+++ packages/mf2-messageformat/src/cst-parser/as-data-model.ts
@@ -1,4 +1,4 @@
-import type * as Model from '../data-model.js';
+import type * as Model from '../data-model/index.js';
 import { MessageSyntaxError } from '../errors.js';
 import type * as CST from './cst-types.js';
