+++ packages/mf2-messageformat/src/data-model/expression/function-annotation.test.ts
@@ -1,4 +1,8 @@
-import { MessageFormat, MessageFunctions, MessageNumberPart } from '../index';
+import {
+  MessageFormat,
+  MessageFunctions,
+  MessageNumberPart
+} from '../../index';
 
 test('Custom function', () => {
   const functions = {
