+++ packages/mf2-messageformat/src/data-model/markup.test.ts
