+++ packages/mf2-messageformat/src/index.ts
@@ -1,13 +1,12 @@
 export { asDataModel, parseMessage, type CST } from './cst-parser/index.js';
-export * from './data-model.js';
+export * from './data-model/index.js';
+export * from './data-model/expression/index.js';
 export {
   MessageDataModelError,
   MessageError,
   MessageResolutionError,
   MessageSyntaxError
 } from './errors.js';
-export * from './expression/index.js';
-export { isMarkup } from './markup/index.js';
 export { MessageFormat, MessageFormatOptions } from './messageformat.js';
 export { stringifyMessage } from './stringifier/message.js';
 export * from './runtime/index.js';
