+++ packages/mf2-messageformat/src/messageformat.ts
@@ -1,9 +1,12 @@
 import { asDataModel, parseMessage } from './cst-parser/index.js';
-import type { Message } from './data-model';
+import type { Message } from './data-model/index.js';
 import { MessageError, MessageResolutionError } from './errors.js';
 import type { Context } from './format-context';
-import { UnresolvedExpression, resolveExpression } from './expression/index.js';
-import { formatMarkup } from './markup/index.js';
+import {
+  UnresolvedExpression,
+  resolveExpression
+} from './data-model/expression/index.js';
+import { formatMarkup } from './data-model/markup.js';
 import {
   MessageFunctions,
   MessagePart,
