+++ packages/mf2-messageformat/src/select-pattern.ts
@@ -1,6 +1,6 @@
-import type { Message, Pattern } from './data-model';
+import type { Message, Pattern } from './data-model/index.js';
 import { MessageSelectionError } from './errors';
-import { resolveExpression } from './expression/index.js';
+import { resolveExpression } from './data-model/expression/index.js';
 import type { Context } from './format-context';
 
 export function selectPattern(
